+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [keep a changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ### Fixed
 - Normalise whitespace (`\s` → ` `) in messages to avoid newlines and tabs breaking in-progress message updating.
+- Assume details message to be preformatted, if it contains newline characters (`\n`). Preformatted message details are wrapped so that newline characters are maintained. This makes, for example, stack traces and console output in message details more readable.
 
 ## [v1.0.1] - 2022-08-30